+++ CHANGELOG.md
@@ -15,6 +15,7 @@ disks, etc. to intermediate storage like a chest.
 - The Wrench now dismantles devices when crouching.
     - The Disk Drive in item form now supports rendering of disks that were dismantled.
     - In order to retain Controller energy, the Controller must now be dismantled.
+- You can now use any Wrench from other mods in order to rotate or dismantle.
 - Added item storage blocks.
 
 ### Fixed